Variable handling
Haanga2 aims to support offline compilation. In order to do this we needed to fix an inherited problem, PHP treats arrays and object with differents notations, therefore we need to guess if a given variable is an array or object. In order to make it as less error prone as possible Haanga 2 introduces variables modifiers.

1. Variables which starts with T_AT (@) will be treated as arrays:

@foo.@bar.foo => $foo['bar']['foo'];

2. Variable which starts with T_DOLLAR ($) will be treated as object:

$foo["foo"] => $foo->{"foo"};
$foo[foo] => $foo->{$foo};
$foo[$foo].bar => $foo->{$foo}->bar;

3. If there are no modifiers we try to guess the proper types checking the contexts variables

4. If everything else from above fails `[]` generates arrays, and `.` objects.